Improve Link Test
